% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateShiftBayesFactor.R
\name{calculateShiftBayesFactor}
\alias{calculateShiftBayesFactor}
\title{Bayes Factors in support of a shift in diversification rates over a given time interval.}
\usage{
calculateShiftBayesFactor(
  rate.trace,
  time.trace,
  rate.name,
  time.name,
  t1,
  t2,
  prior.prob = 0.5,
  decrease = TRUE,
  return.2lnBF = TRUE
)
}
\arguments{
\item{rate.trace}{(list; no default) The processed Rev output of the rate of interest throught time for computation (output of readTrace()).}

\item{time.trace}{(list; no default) The processed Rev output of the change/interval times of the rate of interest throught time for computation (output of readTrace()).}

\item{rate.name}{(character; no default) The name of the parameter (e.g. "speciation") for which Bayes Factor is to be calculated.}

\item{time.name}{(character; no default) The name of the interval times (e.g. "interval_times) for the rate change times.}

\item{t1}{(numeric; no default) Support will be assesed for a shift between time t1 and time t2 (t1 < t2).}

\item{t2}{(numeric; no default) Support will be assesed for a shift between time t1 and time t2 (t1 < t2).}

\item{prior.prob}{(numeric; 0.5) The prior probability of a shift over this interval (default of 0.5 applies to standard HSMRF- and GMRF-based models).}

\item{decrease}{(logical; default TRUE) Should support be assessed for a decrease in the parameter (if TRUE) or an increase (if FALSE) between t1 and t2?}

\item{return.2lnBF}{(logical; TRUE) Should the 2ln(BF) be returned (if TRUE) or simply the BF (if FALSE)?}
}
\value{
The Bayes Factor.
}
\description{
This function computes the Bayes Factor in favor of a rate-shift between time t1 and t2 (t1 < t2).
The default assumption (suitable to standard HSMRF and GMRF models) is that the prior probability of a shift is 0.5.
}
\examples{
\dontrun{

speciation_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_times.p", package="RevGadgets")
speciation_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_rates.p", package="RevGadgets")

speciation_rate <- readTrace(speciation_rate_file,burnin = 0.25)
speciation_times <- readTrace(speciation_time_file,burnin = 0.25)

calculateShiftBayesFactor(speciation_rate,
                          speciation_times,
                          "speciation","
                          interval_times",
                          0.0,40.0,
                          decrease=FALSE)
}

}
\references{
Kass and Raftery (1995) Bayes Factors.
\emph{JASA}, \bold{90 (430)}, 773-795.
}
