% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colFun.R
\name{colFun}
\alias{colFun}
\title{Color Function}
\usage{
colFun(n)
}
\arguments{
\item{n}{(integer; no default) Number of colors to return. Maximum of 12.}
}
\value{
Character vector of color hex codes.
}
\description{
Produce default RevGadgets colors
}
\details{
Produces a vector of colors from the default RevGadgets colors
of length given by n, maximum of 12 colors.
}
\examples{

my_colors <- colFun(2)

}
