% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAP.R
\name{getMAP}
\alias{getMAP}
\title{get MAP}
\usage{
getMAP(var)
}
\arguments{
\item{var}{(numeric vector; no default) Vector of the samples from the
trace of a quantitative variable}
}
\value{
the MAP estimate
}
\description{
Calculates the Maximum a Posteriori estimate for the trace of a
quantitative variable
}
\details{
Uses the SANN method of the optim() function to approximate the MAP estimate
}
\examples{
\dontrun{

file <- system.file("extdata",
    "sub_models/primates_cytb_GTR.p", package="RevGadgets")
trace <- readTrace(paths = file)
MAP <- getMAP(trace[[1]]$"pi[1]")
}

}
