% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAncStateDiscretePie.R
\name{plotAncStateDiscretePie}
\alias{plotAncStateDiscretePie}
\title{Plot ancestral states (discrete, pie)}
\usage{
plotAncStateDiscretePie(
  t,
  cladogenetic = T,
  tip_pie_diameter = 1,
  node_pie_diameter = 1,
  shoulder_pie_diameter = 1,
  tip_labels_remove_underscore = T,
  tip_labels_italics = F,
  pie_nudge_x = 0,
  pie_nudge_y = 0,
  show_tip_labels = F,
  tip_label_size = 5,
  tip_label_offset = 0,
  alpha = 0.5,
  state_labels,
  state_colors,
  collapse_states = T,
  show_legend = T
)
}
\description{
[ function tags to be written ]
}
