% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAncStatesDiscrete.R
\name{plotAncStatesDiscrete}
\alias{plotAncStatesDiscrete}
\title{Plot discrete ancestral states on tree}
\usage{
plotAncStatesDiscrete(
  t,
  summary_statistic,
  tree_layout = "rectangular",
  include_start_states = FALSE,
  xlim_visible = c(0, 40),
  ylim_visible = NULL,
  tip_label_size = 4,
  tip_label_offset = 5,
  tip_label_italics = FALSE,
  tip_node_size = 2,
  tip_node_shape = 15,
  node_label_size = 4,
  node_pp_label_size = 0,
  node_label_nudge_x = 0.1,
  node_pp_label_nudge_x = 0.1,
  shoulder_label_size = 3,
  shoulder_label_nudge_x = -0.1,
  node_pie_diameter = 1.1,
  tip_pie_diameter = 1.08,
  pie_nudge_x = 0,
  pie_nudge_y = 0,
  alpha = 0.5,
  node_size_range = c(6, 15),
  color_low = "#D55E00",
  color_mid = "#F0E442",
  color_high = "#009E73",
  show_state_legend = TRUE,
  show_posterior_legend = TRUE,
  show_tree_scale = TRUE,
  state_labels = NULL,
  state_colors = NULL,
  title = "",
  fig_height = 7,
  fig_width = 7,
  ...
)
}
\arguments{
\item{t}{(treedata; no default) The processed output from processAncStatesDiscrete}

\item{summary_statistic}{(character; no default) The}

\item{tree_layout}{(character; "rectangular) How tree is represented (see ggtree doc).}

\item{include_start_states}{(logical; FALSE) Whether start states for each anagenetic branch is to be plotted on the shoulders. Only applicable for cladogenetic models where state changes may occur at nodes}

\item{xlim_visible}{(numeric; NULL) Plotting limits of the x-axis in cartesian space. If specified, needs to be a vector of length 2.}

\item{ylim_visible}{(numeric; NULL) Plotting limits of the y-axis in cartesian space. If specified, needs to be a vector of length 2.}

\item{tip_label_size}{(numeric; 4) Font size of tip labels}

\item{tip_label_offset}{(numeric; 5) Horizontal offset of tip labels from end of tips}

\item{tip_label_italics}{(logical; FALSE) Whether or not tip labels should be italicized}

\item{tip_node_size}{(numeric; 2) Size of tip points; parsed to geom_tippoint()}

\item{tip_node_shape}{(numeric; 15) Shape of tip points; parsed to geom_tippoint()}

\item{node_label_size}{(numeric; 4) Font size of node labels}

\item{node_pp_label_size}{(numeric; 0) Font size of node labels for posterior probabilities}

\item{node_label_nudge_x}{(numeric; 0.1) Degree of horizontal offset of node labels.}

\item{node_pp_label_nudge_x}{(numeric; 0.1) Degree of horizontal offet of node posterior probability labels}

\item{shoulder_label_size}{(numeric; 3) Font size of labels at node shoulders}

\item{shoulder_label_nudge_x}{(numeric; 3) Degree of horizontal offset of labels at node shoulders}

\item{node_pie_diameter}{(numeric; 1.10) Size of pies at nodes (shoulders pies are 0.9x in size)}

\item{tip_pie_diameter}{(numeric; 1.08) Size of pie at tips}

\item{pie_nudge_x}{(numeric; 0.0) Degree of horizontal offset of pie charts}

\item{pie_nudge_y}{(numeric; 0.0) Degree of vertical offset of pie charts}

\item{alpha}{(numeric; 0.5) Degree of transparency in pies and points}

\item{node_size_range}{(numeric; c(6,15)) Size range of points (used in plotMAP)}

\item{color_low}{(character;"#D55E00")}

\item{color_mid}{(character; "#F0E442")}

\item{color_high}{(character; "#009E73")}

\item{show_state_legend}{(logical, TRUE),}

\item{show_posterior_legend}{(logical, TRUE)}

\item{show_tree_scale}{(logical, TRUE)}

\item{state_colors}{(character, NULL) Vector of colours to be used for states. Must be of equal length to the number of states.}

\item{title}{(character, NULL) Title of plot}

\item{...}{(various)}
}
\value{

}
\description{
Function to plot ancestral states and the associated uncertainty
for continuous and discrete characters. Based on ggtree plotting
functionality
}
\details{
For discrete characters 'summary_statistic="MAP"' should be used,
and for continuous characters 'summary_statistic="mean"'. If the
tree and tip labels do not fit in the screen, adjust the visible
area using the 'xlim_visible' argument.

If 'summary_statistic="MAP"', the maximum a posteriori ancestral
state will be plotted on each node. The color corresponds to the
character state, and the size of the circle represents the posterior
probability of that state. Cladogenetic models that estimate
ancestral states for both the beginning and end of each branch
are plotted by setting "include_start_states=TRUE".

Maximum a posteriori ancestral chromosome numbers can be plotted
with 'summary_statistic="MAPChromosome"'. For chromosomes the
color represents the posterior probability and the size of the
circle represents the chromosome number.

For 'summary_statistic="mean"' the color represents the size of
the 95% confidence interval, and the size of the cirlce represents
represents the mean character state.
}
