% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivRates.R
\name{plotDivRates2}
\alias{plotDivRates2}
\title{Plot Diversification Rates}
\usage{
plotDivRates2(plotdata)
}
\arguments{
\item{plotdata}{(data frame) A table consisting of the following columns:
\itemize{
\item{mean - the mean of the variable},
\item{lower - the lower bounds of the credibility interval}
\item{upper - the upper bounds of the credibility interval}
\item{time - the time units on the x-axis}
\item{item - the variable name, e.g. "speciation rate" or "extinction rate"}
}}
}
\value{
A ggplot object
}
\description{
Plots the output of an episodic diversification rate analysis
}
\details{
Plots the output of episodic diversification rate analyses. Takes as input the
output of processDivRates() and plotting parameters. For now, only variable names
(under "item") that contain the word "rate" are included in the plot.

The return object can be manipulated. For example, you can change the axis labels,
the color palette, whether the axes are to be linked, or the overall plotting style/theme,
just as with any ggplot object.
}
\examples{

\dontrun{
# first run processDivRates()
speciation_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_times.p", package="RevGadgets")
speciation_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_rates.p", package="RevGadgets")
extinction_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_times.p", package="RevGadgets")
extinction_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_rates.p", package="RevGadgets")

primates <- processDivRates(speciation_time_log = speciation_time_file,
                            speciation_rate_log = speciation_rate_file,
                            extinction_time_log = extinction_time_file,
                            extinction_rate_log = extinction_rate_file,
                            burnin = 0.25)

# generate the plot data
plotdata <- makePlotData(primates, probs = c(0.05, 0.95)) # Specify credibility interval for the ribbon plot

# then plot results:
p <- plotDivRates2(plotdata);p

# let's say we want to change the x-axis
p <- p + xlab("Millions of years ago");p

# let's say we want to make the time intervals visible,
# to emphasize that the model estimates rates in bins
p <- p + geom_point();p

# let's say we don't want to plot relative-extinction rate,
# and use the same y-axis for all three rates
plotdata2 <- plotdata \%>\%
               filter(item != "relative-extinction rate")
p2 <- plotDivRates2(plotdata2)
p2 <- p2 +
       facet_wrap(vars(item),
       scale = "fixed");p2
}

}
