% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHiSSE.R
\name{plotHiSSE}
\alias{plotHiSSE}
\title{plotHiSSE}
\usage{
plotHiSSE(rates)
}
\arguments{
\item{rates}{(data.frame; no default) a data frame containing columns
"value", "rate", "hidden_state", "observed_state" (such as the output
of processSSE())}
}
\value{
a ggplot object
}
\description{
plotHiSSE
}
\examples{
\donttest{
hisse_file <- system.file("extdata",
                          "sse/primates_HiSSE_2.p",
                          package="RevGadgets")
pdata <- processSSE(hisse_file)
p <- plotHiSSE(pdata);p

# change colors:
p + ggplot2::scale_fill_manual(values = c("red","green"))

# change x-axis label
p + ggplot2::xlab("Rate (events/Ma)")
}
}
