% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMAP.R
\name{plotMAP}
\alias{plotMAP}
\title{plot MAP}
\usage{
plotMAP(
  t,
  cladogenetic = FALSE,
  tip_labels = TRUE,
  tip_labels_size = 2,
  tip_labels_offset = 1,
  tip_labels_italics = FALSE,
  tip_labels_remove_underscore = TRUE,
  tip_labels_states = FALSE,
  tip_labels_states_size = 2,
  tip_labels_states_offset = 0.1,
  node_labels_as = NULL,
  node_labels_size = 2,
  node_labels_offset = 0.1,
  node_size_as = "state_posterior",
  node_color_as = "state",
  node_shape_as = NULL,
  node_shape = 19,
  node_color = "default",
  node_size = c(2, 6),
  tip_states = TRUE,
  tip_states_size = node_size,
  tip_states_shape = node_shape,
  state_transparency = 0.75,
  show_state_legend = TRUE,
  show_posterior_legend = TRUE,
  tree_layout = "rectangular"
)
}
\arguments{
\item{t}{(treedata object; none) Output of processAncStatesDiscrete() function
containing tree and ancestral states.}

\item{cladogenetic}{(logical; FALSE) Plot shoulder states of cladogenetic analyses?}

\item{tip_labels}{(logical; TRUE) Label taxa labels at tips?}

\item{tip_labels_size}{(numeric; 2) Size of tip labels.}

\item{tip_labels_offset}{(numeric; 1) Horizontal offset of tip labels from tree.}

\item{tip_labels_italics}{(logical; FALSE) Italicize tip labels?}

\item{tip_labels_remove_underscore}{(logical; TRUE) Remove underscores from tip labels?}

\item{tip_labels_states}{(logical; FALSE) Optional plotting of text at tips in addition
to taxa labels.}

\item{tip_labels_states_size}{(numeric; 2) Size of state labels at tips. Ignored if
tip_labels_states is FALSE.}

\item{tip_labels_states_offset}{(numeric; 0.1) Horizontal offset of tip state labels.
Ignored if tip_labels_states = NULL.}

\item{node_labels_as}{(character; NULL) Optional plotting of text at nodes. Possible
values are "state" for the ancestral states , "state_posterior" for posterior probabilities
of the estimated ancestral state, "node_posterior" or the posterior probability of the node on the tree,
or NULL for not plotting any text at the nodes (default).}

\item{node_labels_size}{(numeric; 2) Size of node labels text. Ignored if node_labels_as = NULL.}

\item{node_labels_offset}{(numeric; 0.1) Horizontal offset of node labels from nodes.
Ignored if node_labels_as = NULL.}

\item{node_size_as}{(character; "state_posterior") How to vary size of node symbols. Options
are "state_posterior" (default) for posterior probabilities of the estimated ancestral state,
"node_posterior" or the posterior probability of the node on the tree, "state" for vary size by the
ancestral state itself in cases where there are many character states (e.g. chromosome numbers;
we do not recommend this option for characters with few states), or NULL for fixed symbol size.}

\item{node_color_as}{(character; "state") How to vary to color of node symbols. Options are
"state" (default) to vary by estimated ancestral states, "state_posterior" for posterior probabilities
of the estimated ancestral state, "node_posterior" or the posterior probability of the node on the tree,
or NULL to set all as one color.}

\item{node_shape_as}{(character; NULL) Option to vary node symbol by shape. Options are NULL
to keep shape constant or "state" to vary shape by ancestral state.}

\item{node_shape}{(integer; 19) Shape type for nodes. If node_shape_as = "state", provide a vector
with length of the number of states.}

\item{node_color}{("character"; "default") Colors for node symbols. Defaults to default RevGadgets
colors. If node_color_as = "state', provide a vector of length of the character states.
If node_color_as = "posterior", provide a vector of length 2 to generate a color gradient.}

\item{node_size}{(numeric; c(2, 6)) Range of sizes, or fixed size, for node symbols.
If node_size_as = "state_posterior", "node_posterior", or "state", numeric vector of length two.
If node_size_as = NULL, numeric vector of length one.}

\item{tip_states}{(logical; TRUE) Plot states of taxa at tips?}

\item{tip_states_size}{(numeric; node_size) Size for tip symbols. Defaults to the same
size as node symbols.}

\item{tip_states_shape}{(integer; node_shape) Shape for tip symbols. Defaults to the same
as node symbols.}

\item{state_transparency}{(integer; 0.75) Alpha (transparency) of state symbols- varies from
0 to 1.}

\item{show_state_legend}{(logical; TRUE) Plot legend for states?}

\item{show_posterior_legend}{(logical; TRUE) Plot legend for posterior probabilities?}

\item{tree_layout}{(character; "rectangular") Tree shape layout, passed to ggtree(). Options
are 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle', or 'daylight'}
}
\description{
Plot the MAP character state and posterior probabilities as points on nodes (size = pp, colour = state)
and character state on tips (different shape from nodes, colour = state)
}
\examples{

\dontrun{
# Standard ancestral state reconstruction example with various aesthetics

# process file
file <- system.file("extdata", "comp_method_disc/ase_freeK.tree", package="RevGadgets")
example <- processAncStatesDiscrete(file, state_labels = c("1" = "Awesome", "2" = "Beautiful", "3" = "Cool!"))

# have states vary by color and indicate state pp with size (default)
plotMAP(t = example, tip_labels_italics = T)

# have states vary by color and symbol, label nodes with pp of states
plotMAP(t = example,  node_shape_as = "state", node_size = 4, node_shape = c(15, 17,20),
        node_size_as = NULL, node_labels_as = "state_posterior")

# black and white figure - state as symbols and state pp with text
plotMAP(t = example, node_color_as = NULL, node_shape_as = "state", node_shape =  c(15, 17,20),
        node_size_as = NULL, node_size = 4, node_labels_as = "state_posterior",
        node_color = "grey", state_transparency = 1)

# default with circular tree
plotMAP(t = example, tree_layout = "circular")

# Chromosome evolution example

# process file
file <- system.file("extdata", "chromo/ChromEvol_simple_final.tree", package="RevGadgets")
chromo_example <- processAncStatesDiscrete(file, labels_as_numbers = TRUE)

# plot
plotMAP(t = chromo_example, node_color_as = "state_posterior", node_size_as = "state",
        node_color = RevGadgets:::.colFun(2), tip_labels_offset = 0.005,
        node_labels_as = "state", node_labels_offset = 0, tip_labels_states = TRUE,
        tip_labels_states_offset = 0, tip_states = FALSE)

# DEC example (cladogenetic)

# process file
file <- system.file("extdata", "dec/simple.ase.tre", package="RevGadgets")
labs <- c("1" = "K", "2" = "O", "3" = "M", "4" = "H", "5" = "KO",
"6" = "KM", "7" = "OM", "8" = "KH", "9" = "OH", "10" = "MH", "11" = "KOM",
"12" = "KOH", "13" = "KMH", "14" = "OMH", "15" = "KOMH")
dec_example <- processAncStatesDiscrete(file, state_labels = labs)

# plot
plotMAP(t = dec_example, cladogenetic = TRUE, tip_labels_offset = 0.5)
}

}
