% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSSE.R
\name{plotMuSSE}
\alias{plotMuSSE}
\title{plotMuSSE}
\usage{
plotMuSSE(x)
}
\arguments{
\item{x}{a data frame containing columns "value", "rate", "hidden_state", "observed_state"}
}
\value{
a ggplot object#'
}
\description{
plotMuSSE
}
\examples{
bisse_file <- system.file("extdata", "sse/primates_BiSSE_activity_period.log", package="RevGadgets")

tr <- readTrace(bisse_file)[[1]]

pdata <- processSSE(tr)
p <- plotMuSSE(pdata)

}
