% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPie.R
\name{plotPie}
\alias{plotPie}
\title{plot Pie}
\usage{
plotPie(
  t,
  cladogenetic = FALSE,
  tip_labels = TRUE,
  tip_labels_size = 2,
  tip_labels_offset = 1,
  tip_labels_italics = FALSE,
  tip_labels_remove_underscore = TRUE,
  tip_labels_states = FALSE,
  tip_labels_states_size = 2,
  tip_labels_states_offset = 0.1,
  node_labels_as = NULL,
  node_labels_size = 2,
  node_labels_offset = 0.1,
  pie_colors = "default",
  node_pie_size = 2,
  shoulder_pie_size = node_pie_size,
  tip_pies = TRUE,
  tip_pie_size = 1,
  node_pie_nudge_x = 0,
  node_pie_nudge_y = 0,
  tip_pie_nudge_x = node_pie_nudge_x,
  tip_pie_nudge_y = node_pie_nudge_y,
  shoulder_pie_nudge_x = node_pie_nudge_x,
  shoulder_pie_nudge_y = node_pie_nudge_y,
  collapse_states = FALSE,
  collapse_states_threshold = 0.05,
  state_transparency = 0.75
)
}
\arguments{
\item{t}{(treedata object; none) Output of processAncStatesDiscrete() function
containing tree and ancestral states.}

\item{cladogenetic}{(logical; FALSE) Plot shoulder states of cladogenetic analyses?}

\item{tip_labels}{(logical; TRUE) Label taxa labels at tips?}

\item{tip_labels_size}{(numeric; 2) Size of tip labels.}

\item{tip_labels_offset}{(numeric; 1) Horizontal offset of tip labels from tree.}

\item{tip_labels_italics}{(logical; FALSE) Italicize tip labels?}

\item{tip_labels_remove_underscore}{(logical; TRUE) Remove underscores from tip labels?}

\item{tip_labels_states}{(logical; FALSE) Optional plotting of text at tips in addition
to taxa labels. Will plot the MAP state label.}

\item{tip_labels_states_size}{(numeric; 2) Size of state labels at tips. Ignored if
tip_labels_states is FALSE.}

\item{tip_labels_states_offset}{(numeric; 0.1) Horizontal offset of tip state labels.
Ignored if tip_labels_states = NULL.}

\item{node_labels_as}{(character; NULL) Optional plotting of text at nodes. Possible
values are "state" for the MAP ancestral states, "state_posterior" for posterior probabilities
of the estimated ancestral state, "node_posterior" or the posterior probability of the node on the tree,
or NULL for not plotting any text at the nodes (default).}

\item{node_labels_size}{(numeric; 2) Size of node labels text. Ignored if node_labels_as = NULL.}

\item{node_labels_offset}{(numeric; 0.1) Horizontal offset of node labels from nodes.
Ignored if node_labels_as = NULL.}

\item{state_transparency}{(integer; 0.75) Alpha (transparency) of state symbols- varies from
0 to 1.
#'}

\item{node_pie_size_as}{(character; "posterior") How to vary size of node symbols. Options
are "node_posterior" or the posterior probability of the node on the tree or NULL for fixed symbol size.}

\item{node_color_as}{(character; "state") How to vary to color of node symbols. Options are
"state" (default) to vary by estimated ancestral states, "state_posterior" for posterior probabilities
of the estimated ancestral state, "node_posterior" or the posterior probability of the node on the tree,
or NULL to set all as one color.}

\item{node_color}{("character"; "default") Colors for node symbols. Defaults to default RevGadgets
colors. If node_color_as = "state', provide a vector of length of the character states.
If node_color_as = "posterior", provide a vector of length 2 to generate a color gradient.}

\item{node_size}{(numeric; c(2, 6)) Range of sizes, or fixed size, for node symbols.
If node_size_as = "state_posterior", "node_posterior", or "state", numeric vector of length two.
If node_size_as = NULL, numeric vector of length one.}

\item{tip_states}{(logical; TRUE) Plot states of taxa at tips?}

\item{tip_states_size}{(numeric; node_size) Size for tip symbols. Defaults to the same
size as node symbols.}

\item{tip_states_shape}{(integer; node_shape) Shape for tip symbols. Defaults to the same
as node symbols.}
}
\description{
Plot character states and posterior probabilities as pies on nodes.
}
