% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivRates.R
\name{plotPopulationSize}
\alias{plotPopulationSize}
\title{Title}
\usage{
plotPopulationSize(df, plot_var = "size")
}
\arguments{
\item{df}{a data frame, such as produced by processDivRates(), containing 
the data on rates and interval times for each type of rate to be 
plotted (speciation rate, etc.).}

\item{plot_var}{only include variable that contain the string, default "rate" in the name}
}
\value{
a ggplot object
}
\description{
Title
}
\examples{
library(tibble)

df <- tibble("time" = c(0.0, 1.0, 2.0, 3.0, 4.0),
             "time_end" = c(1.0, 2.0, 3.0, 4.0, 5.0),
             "value" = c(1.0, 1.5, 2.0, 1.5, 1.5),
             "upper" = c(3.5, 7.0, 6.5, 5.0, 5.0),
             "lower" = c(0.5, 0.0, 0.5, 0.5, 0.8),
             "item" = "population size")
             
plotPopulationSize(df)
}
