% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPostPredStats.R
\name{plotPostPredStats}
\alias{plotPostPredStats}
\title{plot Posterior Predictive Statistics}
\usage{
plotPostPredStats(data)
}
\arguments{
\item{data}{(list of data frames; no default) A list of data frames
of the empirical and simulated values, such as the output of
processPostPredStats.R}
}
\value{
A list of ggplot objects, where each plot contains a density distribution
of the predicted values and a dashed line of the empirical value. The light blue
shaded region of the density plot corresponds to the 5% two-sided quantile and
the dark blue corresponds to the 2% two-sided quantile.
}
\description{
Plots the posterior predictive statistics data
}
\details{
Produces one ggplot object per metric. Intended
to plot the results of the RevBayes tutorial:
Assessing Phylogenetic Reliability Using RevBayes and P3
Model adequacy testing using posterior prediction (Data Version).
}
\examples{

\dontrun{
file_sim <- system.file("extdata",
    "PPS/simulated_data_pps_example.csv", package="RevGadgets")
file_emp <- system.file("extdata",
    "PPS/empirical_data_pps_example.csv", package="RevGadgets")
t <- processPostPredStats(path_sim = file_sim,
                         path_emp = file_emp)
plots <- plotPostPredStats(data = t)

}

}
