% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{Plot tree}
\usage{
plotTree(tree, fossils = NULL)
}
\arguments{
\item{tree}{(list of lists of tidytree objects; no default) Name of a list of lists of
tidytree objects, such as produced by readTrees(). This object should only contain
only one summary tree from one trace file. If it contains multiple trees or multiple
traces, only the first will be used.}

\item{fossils}{(character vector; NULL) Optional vector of the tips that are fossils,
to be sure that these tips are highlighted in red.}
}
\value{
returns a single plot object.
}
\description{
Plots a single tree, such as an MCC or MAP tree.
}
\details{
Plots a single tree, such as an MCC or MAP tree, with
optionally labele posterior probabilities at nodes, a
timescale plotted on the x - axis, and 95\% CI for node ages.
}
\examples{

}
