% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValidation.R
\name{plotValidation}
\alias{plotValidation}
\title{Plot Validation Results from RevBayes Analyses}
\usage{
plotValidation(validation_results, parameter = NULL)
}
\arguments{
\item{validation_results}{(list; no default) A named list of data frames as returned by the
`processValidation` function, where each data frame corresponds to a parameter and contains
the HPD width, the count of simulations where the true value was within the HPD interval,
the total count and the coverage frequency.}

\item{parameter}{(character string; optional) The name of a single parameter to plot from the
'validation_results' list. If 'NULL' (the default), the function will generate a list of plots for all
parameters found in the input.}
}
\value{
A ggplot object if a single parameter is specified, or a named list of ggplot objects
  if multiple parameters are present in the input list.
}
\description{
Creates a ggplot visualization of RevBayes validation results. It plots the observed coverage
probability against the expected probability (the HPD width). A
perfectly calibrated model should yield a diagonal line from (0,0) to (1,1).
}
\examples{
\dontrun{
# First, generate the validation results
validation_output <- processValidation(path = "output_my_analysis", n_reps = 100)

# Plot the results for a single parameter
p_alpha <- plotValidation(validation_output, parameter = "alpha")

# Display the plot
p_alpha

# Customize and save it
p_alpha <- p_alpha + ggplot2::labs(title = "Validation for Alpha Parameter")
ggplot2::ggsave("alpha_validation.png", p_alpha, width = 5, height = 5)
}

}
