% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorSamplesToParametricPrior.R
\name{posteriorSamplesToParametricPrior}
\alias{posteriorSamplesToParametricPrior}
\title{Priors from MCMC samples}
\usage{
posteriorSamplesToParametricPrior(
  samples,
  distribution,
  variance_inflation_factor = 2
)
}
\arguments{
\item{samples}{(numeric vector; no default) MCMC samples for a single
parameter.}

\item{distribution}{(character; no default) The distribution to fit. Options
are gamma for strictly positive
parameters and normal for unbounded parameters.}

\item{variance_inflation_factor}{(single numeric value; default = 2.0) Makes
the prior variance larger than the variance of the posterior}
}
\value{
Numeric vector of parameters with names (to avoid rate/scale and
var/sd confusion).
}
\description{
Turn posterior samples collected by MCMC into a parametric prior
distribution.
}
\details{
The distributions are fit by the method of moments.
The function allows inflating the prior variance relative to the posterior
being supplied.
}
\examples{

\dontrun{

speciation_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_times.p", package="RevGadgets")
speciation_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_rates.p", package="RevGadgets")
extinction_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_times.p", package="RevGadgets")
extinction_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_rates.p", package="RevGadgets")

primates <- processDivRates(speciation_time_log = speciation_time_file,
                            speciation_rate_log = speciation_rate_file,
                            extinction_time_log = extinction_time_file,
                            extinction_rate_log = extinction_rate_file,
                            burnin = 0.25)

speciation_rates <-
       dplyr::pull(primates[which(primates$item == "speciation rate"),],
                  "mean")
speciation_1_gamma_prior <-
          posteriorSamplesToParametricPrior(speciation_rates,"gamma")

}

}
