% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAncStatesDiscrete.R
\name{processAncStatesDiscrete}
\alias{processAncStatesDiscrete}
\title{Process ancestral states (discrete)}
\usage{
processAncStatesDiscrete(path, state_labels = NULL)
}
\description{
[ function tags to be written ]
}
