% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBranchData.R
\name{processBranchData}
\alias{processBranchData}
\title{processBranchData}
\usage{
processBranchData(
  tree,
  df,
  burnin = 0.25,
  parnames = c("avg_lambda", "avg_mu", "num_shifts"),
  summary = "median"
)
}
\arguments{
\item{tree}{a phylogenetic tree in the treedata format}

\item{df}{data frame with branch specific data}

\item{burnin}{fraction of the markov-chain to discard}

\item{parnames}{defaults to c("avg_lambda", "avg_mu", "num_shifts")}

\item{summary}{function to summarize the continuous parameter. Typically mean or median}
}
\value{
a treedata file with attached branch-specific data
}
\description{
processBranchData
}
