% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDivRates.R
\name{processDivRates}
\alias{processDivRates}
\title{Process Diversification Rates}
\usage{
processDivRates(speciation_time_log = "", speciation_rate_log = "",
  extinction_time_log = "", extinction_rate_log = "", burnin = 0.25)
}
\arguments{
\item{speciation_time_log}{(vector of character strings or single character string; "") Path to speciation times log file(s)}

\item{speciation_rate_log}{(vector of character strings or single character string; "") Path to speciation rates log file(s)}

\item{extinction_time_log}{(vector of character strings or single character string; "") Path to extinction times log file(s)}

\item{extinction_rate_log}{(vector of character strings or single character string; "") Path to extinction rates log file(s)}
}
\value{
List object with processed rate and time parameters.
}
\description{
Processing the output of a episodic diversification rate analysis with mass-extinction events.
}
\details{
For processing the output of an episodic diversiation rate analysis. processDivRates()
assumes that the epochs are fixed rather than inferred. Additionally, it assumes
that times correspond to rates such that the first rate parameter (i.e. speciation[1])
corresponds to the present. Conversely, the first time parameter
(i.e. interval_times[1]) corresponds to the first time interval after the present,
moving backwards in time. processDivRates() relies on readTrace and produces a list
object that can be read by plotDivRates() to vizualize the results. For now,
only one log file per parameter type is accepted (i.e. log files from multipe runs
must be combined before reading into the function).
}
