% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPopSizes.R
\name{processPopSizes}
\alias{processPopSizes}
\title{Process Population Sizes}
\usage{
processPopSizes(
  population_size_log = "",
  interval_change_points_log = "",
  method = "events",
  burnin = 0.25,
  probs = c(0.025, 0.975),
  summary = "mean",
  num_grid_points = 100
)
}
\arguments{
\item{population_size_log}{(vector of character strings or
single character string; "") Path to population sizes log file(s)}

\item{interval_change_points_log}{(vector of character strings or
single character string; "") Path to interval change points log file(s)}

\item{method}{which method was chosen for the analysis, "events" - event-based coalescent process,
"specified" - coalescent process with user-defined interval times, "constant" - constant coalescent process}

\item{burnin}{(single numeric value; default = 0) Fraction of generations to
discard (if value provided is between 0 and 1) or number of generations (if
value provided is greater than 1).}

\item{probs}{(numeric vector; c(0.025, 0.975)) a vector of length two
containing the upper and lower bounds for the confidence intervals.}

\item{summary}{typically "mean" or "median"; the metric to summarize the
posterior distribution. Defaults to "mean"}

\item{num_grid_points}{only if method = "events, defines the number of grid points through time for which to 
evaluate the demographic functions}
}
\value{
List object with processed rate and, if applicable, time parameters.
}
\description{
Processing the output of a coalescent demographic analysis.
}
\details{
For processing the output of a coalescent demographic analysis.
processPopSizes() can process output from different analyses,
either an event-based skyline analysis (method = "events"),
or an analysis with user-defined intervals (method = "specified"),
or an analysis with only one population size and no interval times
(method = "constant").
processPopSizes() assumes that the the first size parameter (i.e. population_size[1])
corresponds to the present. processPopSizes() partly
relies on readTrace and produces a list object that can be read by
plotPopSizes() to visualize the results. For now, only one log file per
parameter type is accepted (i.e. log files from multiple runs must be
combined before reading into the function).
}
