% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPostPredStats.R
\name{processPostPredStats}
\alias{processPostPredStats}
\title{process Posterior Predictive Statistics}
\usage{
processPostPredStats(path_sim, path_emp)
}
\arguments{
\item{path_sim}{(character string; no default) Path to the .csv file
containing the simulated data results}

\item{path_emp}{(character string; no default) Path to the .csv file
containing the empirical values}
}
\value{
A list of data frames
}
\description{
Reads in and processes posterior-predictive statistics
}
\examples{

\dontrun{
file_sim <- system.file("extdata",
    "PPS/simulated_data_pps_example.csv", package="RevGadgets")
file_emp <- system.file("extdata",
    "PPS/empirical_data_pps_example.csv", package="RevGadgets")
t <- processPostPredStats(path_sim = file_sim,
                                          path_emp = file_emp)
}

}
