% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSSE.R
\name{processSSE}
\alias{processSSE}
\title{Title}
\usage{
processSSE(
  tr,
  speciation = "speciation",
  extinction = "extinction",
  speciation_hidden = "speciation_hidden",
  rates = c(speciation, extinction, "net-diversification")
)
}
\arguments{
\item{tr}{a trace object}

\item{speciation}{revbayes variable name}

\item{extinction}{revbayes variable name}

\item{speciation_hidden}{revbayes variable name}

\item{rates}{names of rates to be included in plot}
}
\value{
a data frame
}
\description{
Title
}
\examples{
bisse_file <- system.file("extdata", "sse/primates_BiSSE_activity_period.log", package="RevGadgets")

tr <- readTrace(bisse_file)[[1]]
pdata <- processSSE(tr)
head(pdata)

}
