% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processValidation.R
\name{processValidation}
\alias{processValidation}
\title{Process RevBayes validation Results}
\usage{
processValidation(
  path,
  results_path = NULL,
  parameters = NULL,
  n_reps = 1000,
  n_bins = 50,
  burnin = 0.25,
  verbose = TRUE
)
}
\arguments{
\item{path}{(character; no default) The path to the main output directory which contains
the individual simulation subdirectories (e.g., "output_<analysis_name>").}

\item{results_path}{(character; optional) The path to a directory where the
processed RDS file for each parameter will be saved. If NULL (the default),
the results will not be saved to disk.}

\item{parameters}{(character vector; optional) A vector of parameter names to process.
If NULL (the default), the function will automatically find all parameters in the
output files.}

\item{n_reps}{(numeric; default: 1000) The total number of simulation replicates.}

\item{n_bins}{(numeric; default 50) The number of bins to use for the HPD widths, ranging
from 0 to 1.}

\item{burnin}{(numeric; default: 0.25) The fraction of samples from the MCMC chain to
discard as burn-in.}

\item{verbose}{(logical; default: TRUE) Whether to print progress bar and summary
messages to the console.}
}
\value{
A named list of data frames. Each data frame corresponds to a parameter and contains
the HPD witdth, the count of simulations where the true value was within the HPD interval,
the total count and the coverage frequency.
}
\description{
Reads and processes the output of a set of simulation replicates to assesss
calibration. For each specified parameter, it calculates the coverage
frequency across a range of Highest Posterior Density (HPD) interval widths.
}
\examples{
\dontrun{
# assuming simulation outputs are in "output_my_analysis"
validation_results <- processValidation(
  path = "output_my_analysis",
  results_path = "results_my_analysis",
  n_reps = 1000,
  parameters = c("alpha", "sigma")
)

# The results for the 'alpha' parameter can be accessed as:
alpha_results <- validation_results$alpha
}

}
