% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndParseJSON.R
\name{readAndParseJSON}
\alias{readAndParseJSON}
\title{Read and Parse JSON lines file
Internal function to read in JSON files}
\usage{
readAndParseJSON(file)
}
\arguments{
\item{file}{(character) File path to trace file.}
}
\value{
Data frame.
}
\description{
Read and Parse JSON lines file
Internal function to read in JSON files
}
