% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnTrees.R
\name{readAnnTrees}
\alias{readAnnTrees}
\title{Read annotated trees}
\usage{
readAnnTrees(path)
}
\arguments{
\item{path}{(single character string; no default) File path to tree(s)}
}
\value{
List of treedata objects ("beastList"), of length one if only one tree provided
}
\description{
Reads in a tree files containing annotated trees
}
\details{
Reads in a tree file in nexus format containing a single or multiple
annotated trees. Relies on TreeIO.
}
\examples{
tree_single <- readAnnTrees(path = "data/comp_method_disc/ase_freeK.tree")
tree_multi <- readAnnTrees(path = "data/nexus_multi_ann.nex")

}
