% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBurnin.R
\name{removeBurnin}
\alias{removeBurnin}
\title{Remove Burnin}
\usage{
removeBurnin(trace, burnin)
}
\arguments{
\item{trace}{(list of data frames; no default) Name of a list of data frames,
such as produced by readTrace().}

\item{burnin}{(single numeric value; 0.1) Fraction of generations to
discard (if value provided is between 0 and 1) or number of generations (if
value provided is greater than 1).}
}
\value{
List of dataframes (of length 1 if only 1 log file provided).
}
\description{
Removes burnin from MCMC trace
}
\details{
Removes burnin from an MCMC trace, such as the output of readTrace(). If
multiple traces are provided, this function will remove the burnin from each.
}
\examples{

file <- system.file("extdata",
    "sub_models/primates_cytb_GTR.p", package="RevGadgets")
one_trace <- readTrace(paths = file)

one_trace_burnin <- removeBurnin(trace = one_trace, burnin = 0.1)

}
