\name{RevGadgets-package}
\alias{RevGadgets-package}
\alias{RevGadgets}
\docType{package}
\title{Process output generated by RevBayes}
\description{Process output generated by RevBayes}
\details{
\tabular{ll}{
Package: \tab RevGadgets\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-9-1\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{Sebastian Hoehna and Will A. Freyman

Maintainer: Sebastian Hoehna <Sebastian.Hoehna@gmail.com>
}
\references{}
\seealso{
\code{ape}
\code{coda}
%\code{\link[ape]{ape}}
}
