% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRFGlobalScaleCalculator.R
\name{setMRFGlobalScaleHyperpriorNShifts}
\alias{setMRFGlobalScaleHyperpriorNShifts}
\title{Sets a global scale parameter for a GMRF or HSMRF model given a prior mean number of effective shifts.}
\usage{
setMRFGlobalScaleHyperpriorNShifts(n.episodes, model,
  prior.n.shifts = log(2), shift.size = 2)
}
\arguments{
\item{n.episodes}{(numeric; no default) The number of episodes in the random field (the parameter vector will be this long).}

\item{model}{(character; no default) What model should the global scale parameter be set for? Options are "GMRF" and "HSMRF" for first-order models (also allowable: "GMRF1" and "HSMRF1") and "GMRF2" and HSMRF2" for second-order models.}

\item{prior.n.shifts}{(numeric; log(2)) The desired prior mean number of shifts.}

\item{shift.size}{(numeric; 2) The magnitude of change that defines an effective shift (measured as a fold-change).}
}
\value{
The hyperprior.
}
\description{
This function finds the global scale parameter value that produces the desired prior mean number of "effective" rate shifts.
Given a specified magnitude for an effective shift, shift.size, an effective shift occurs when two adjacent values are more than shift.size-fold apart from each other.
That is, an effective shift is the event that rate[i+1]/rate[i] > shift.size or rate[i+1]/rate[i] < 1/shift.size.
}
\details{
Finding these values for a HSMRF model can take several seconds for large values of n.episodes because of the required numerical integration.
}
\examples{
# Get global scale for a HSMRF model with 100 episodes.
gs <- setMRFGlobalScaleHyperpriorNShifts(100,"HSMRF")

# Plot a draw from this HSMRF distribution
trajectory <- simulateMRF(n.episodes=100,model="HSMRF",global.scale.hyperprior=gs)
plot(1:100,rev(trajectory),type="l",xlab="time",ylab="speciation rate")

}
\references{
Magee et al. (2019) Locally adaptive Bayesian birth-death model successfully detects slow and rapid rate shifts.
doi: https://doi.org/10.1101/853960
}
