% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTrace.R
\name{summarizeTrace}
\alias{summarizeTrace}
\title{Summarize trace}
\usage{
summarizeTrace(trace, quant_char = "", disc_char = "")
}
\arguments{
\item{trace}{(data frame; no default) Name of a single data frame,
such as one data frame produced by readTrace(). If the readTrace() output
contains multiple traces (such as from multiple runs), you should
combine them into one data frame prior to summarizing or summarize each
trace separately.}

\item{quant_char}{(character vector; default = "") Vector of the names of
quantitative characters to summarize.}

\item{disc_char}{(character vector; default = "") Vector of the names  of
discrete characters to summarize.}
}
\value{
List of data frames of length two. The first element
contains the summaries of quantitative variables (if no quantitative
variables included, element is empty). The second element contains
the summaries of discrete variables (again, empty if none provided).
}
\description{
Summarizes trace file(s) that have been read into memory
}
\details{
Summarizes a trace file for continuous or discrete characters by
computing the mean and 95% credible interval for quantitative
character and the probability for each state of discrete characters.
}
\examples{

\dontrun{
file <- system.file("extdata",
    "sub_models/primates_cytb_covariotide.p", package="RevGadgets")
one_trace <- readTrace(path = file)

# continuous character example
trace_sum <- summarizeTrace(trace = one_trace[[1]]
                            quant_char = c("er[1]","er[2]",
                                           "er[3]","er[4]",
                                           "er[5]","er[6]"))
}

}
