% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBranchData.R
\name{matchNodes}
\alias{matchNodes}
\title{Title}
\usage{
matchNodes(phy)
}
\arguments{
\item{phy}{tree in ape format}
}
\value{
a data frame that translates ape node numbers to RevBayes node numbers
}
\description{
Title
}
