% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNodes.R
\name{matchNodes}
\alias{matchNodes}
\title{match Nodes}
\usage{
matchNodes(phy)
}
\arguments{
\item{phy}{tree in ape format}
}
\value{
a data frame that translates ape node numbers to RevBayes node numbers
}
\description{
match Nodes
}
\examples{
\dontrun{
treefile <- system.file("extdata", "bds/primates.tre", package="RevGadgets")
tree <- readTrees(treefile)
map <- matchNodes(tree[[1]][[1]]@phylo)
}

}
