% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiversityOBDP.R
\name{plotDiversityOBDP}
\alias{plotDiversityOBDP}
\title{Plot Diversity distribution from OBDP analysis}
\usage{
plotDiversityOBDP(
  Kt_mean,
  xlab = "Time",
  ylab = "Number of lineages",
  xticks.n.breaks = 5,
  col.Hidden = "dodgerblue3",
  col.LTT = "gray25",
  col.Total = "forestgreen",
  col.Hidden.interval = "dodgerblue2",
  col.Total.interval = "darkolivegreen4",
  palette.Hidden = c("transparent", "dodgerblue2", "dodgerblue3", "dodgerblue4",
    "black"),
  palette.Total = c("transparent", "green4", "forestgreen", "black"),
  line.size = 0.7,
  interval.line.size = 0.5,
  show.Hidden = TRUE,
  show.LTT = TRUE,
  show.Total = TRUE,
  show.intervals = TRUE,
  show.densities = TRUE,
  show.expectations = TRUE,
  use.interpolate = TRUE
)
}
\arguments{
\item{Kt_mean}{(data.frame; no default) The processed data.frame previously obtained with `process_OBDP_output`.}

\item{xlab}{(character; "Time") The label of the x-axis.}

\item{ylab}{(character; "Number of lineages") The label of the y-axis.}

\item{xticks.n.breaks}{(numeric; 5) An integer guiding the number of major breaks.}

\item{col.Hidden}{(character; "dodgerblue3") The color of the hidden lineages plot line.}

\item{col.LTT}{(character; "gray25") The color of the LTT plot line.}

\item{col.Total}{(character; "forestgreen") The color of the total lineages plot line.}

\item{col.Hidden.interval}{(character; "dodgerblue2") The color of the credible interval lines around the hidden lineages plot.}

\item{col.Total.interval}{(character; "darkolivegreen4") The color of the credible interval lines around the total lineages plot.}

\item{palette.Hidden}{(character; c("transparent", "dodgerblue2", "dodgerblue3", "dodgerblue4", "black")) The palette of the hidden lineages plot distribution.}

\item{palette.Total}{(character; c("transparent", "green4", "forestgreen", "black")) The palette of the total lineages plot distribution.}

\item{line.size}{(numeric; 0.7) The width of the lineage plot line.}

\item{interval.line.size}{(numeric; 0.5) The width of the credible interval.}

\item{show.Hidden}{(boolean; TRUE) Whether to show the plot for hidden lineages.}

\item{show.LTT}{(boolean; TRUE) Whether to show the plot for observed lineages.}

\item{show.Total}{(boolean; TRUE) Whether to show the plot for total lineages.}

\item{show.intervals}{(boolean; TRUE) Whether to show the credible intervals.}

\item{show.densities}{(boolean; TRUE) Whether to show the diversity densities.}

\item{show.expectations}{(boolean; TRUE) Whether to show the diversity expectations.}

\item{use.interpolate}{(boolean; TRUE) Whether to interpolate densities.}
}
\value{
A ggplot object
}
\description{
Plots the probability distribution of the number of lineages through time inferred with the 
Occurrence Birth Death Process
#'
}
