% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivRates.R
\name{plotDivRates}
\alias{plotDivRates}
\title{Plot Diversification Rates}
\usage{
plotDivRates(rates, facet = TRUE)
}
\arguments{
\item{rates}{(list of dataframes; no default) A list of dataframes,
such as produced by processDivRates(), containing the data on rates
and interval times for each type of rate to be plotted (e.g.
speciation rate, etc.).}

\item{facet}{(logical; TRUE) plot rates in separate facets.}
}
\value{
A ggplot object
}
\description{
Plots the output of an episodic diversification rate analysis
}
\details{
Plots the output of episodic diversification rate analyses. Takes as
input the output of processDivRates() and plotting parameters.
For now, only variable names (under "item") that contain the word "rate" are
included in the plot.

The return object can be manipulated. For example, you can change the
axis labels, the color palette, whether the axes are to be linked, or the
overall plotting style/theme, just as with any ggplot object.
}
\examples{

\donttest{
# first run processDivRates()
speciation_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_times.p", package="RevGadgets")
speciation_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_rates.p", package="RevGadgets")
extinction_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_times.p", package="RevGadgets")
extinction_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_rates.p", package="RevGadgets")

rates <- processDivRates(speciation_time_log = speciation_time_file,
                         speciation_rate_log = speciation_rate_file,
                         extinction_time_log = extinction_time_file,
                         extinction_rate_log = extinction_rate_file,
                         burnin = 0.25)

# then plot results:
p <- plotDivRates(rates = rates);p

# change the x-axis
p <- p + ggplot2::xlab("Thousands of years ago");p

# change the colors
p <- p + ggplot2::scale_fill_manual(values = c("red",
                                               "green",
                                               "yellow",
                                               "purple")) +
  ggplot2::scale_color_manual(values = c("red",
                                         "green",
                                         "yellow",
                                         "purple"));p

# let's say we don't want to plot relative-extinction rate,
# and use the same y-axis for all three rates
rates <- rates[!grepl("relative-extinction", rates$item),]
p2 <- plotDivRates(rates)
p2 <- p2 + ggplot2::facet_wrap(ggplot2::vars(item), scale = "fixed");p2
}

}
