% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivRates.R
\name{plotDivRates}
\alias{plotDivRates}
\title{Plot Diversification Rates}
\usage{
plotDivRates(output, fig_types = c("speciation rate", "extinction rate",
  "net-diversification rate", "relative-extinction rate"),
  xlab = "million years ago", col = NULL, col_alpha = "50",
  offset = 0, xaxt = "n", yaxt = "s", ...)
}
\arguments{
\item{output}{(list; no default) The processed output for plotting (output of processDivRates()).}

\item{fig_types}{(character vector, c("speciation rate", "extinction rate", "net-diversification rate", "relative-extinction rate")) Which aspects of the model to visualize. See details for a complete description.}

\item{xlab}{(character; "million years ago") The label of the x-axis.}

\item{col}{(character; NULL) Colors used for printing in hex code. Must be of same length as fig_types.}

\item{col_alpha}{(character; "50") Alpha channel parameter for credible intervals plotting. May range from 00 (completely transparent) to FF (completely opaque).}

\item{offset}{(numeric; 0) Value to offset the x-axis by.}

\item{xaxt}{(character; "n") The type of x-axis to plot. By default, no x-axis is plotted (recommended).}

\item{yaxt}{(character; "s") The type of y-axis to plot.}

\item{...}{Parameters delegated to various plotting functions.}
}
\value{
Plots diversification rates, does not return an object.
}
\description{
Plots the output of a episodic diversification rate analysis
}
\details{
Plots the output of diversification rate analyses. Takes as input the
output of processDivRates() and plotting parameters. Does not return an
object. For now, valid figure types include: "speciation rate",
"extinction rate","net-diversification rate", and "relative-extinction rate".
If colors are not provided (parameter col), the plot defaults to preset colors.
}
\examples{

\dontrun{
# first run processDivRates()
speciation_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_times.p", package="RevGadgets")
speciation_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_speciation_rates.p", package="RevGadgets")
extinction_time_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_times.p", package="RevGadgets")
extinction_rate_file <- system.file("extdata",
    "epi_bd/primates_EBD_extinction_rates.p", package="RevGadgets")

primates <- processDivRates(speciation_time_log = speciation_time_file,
                            speciation_rate_log = speciation_rate_file,
                            extinction_time_log = extinction_time_file,
                            extinction_rate_log = extinction_rate_file,
                            burnin = 0.25)
# then plot results:
plotDivRates(output = primates)

}
}
