% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHiSSE.R
\name{plotHiSSE}
\alias{plotHiSSE}
\title{plotHiSSE}
\usage{
plotHiSSE(rates)
}
\arguments{
\item{rates}{(data.frame; no default) a data frame containing columns "value", "rate",
"hidden_state", "observed_state" (such as the output of processSSE())}
}
\value{
a ggplot object
}
\description{
plotHiSSE
}
\examples{
\dontrun{
hisse_file <- system.file("extdata", "sse/primates_HiSSE_2.p", package="RevGadgets")
pdata <- processSSE(hisse_file)
p <- plotHiSSE(pdata);p

# change colors:
library(ggplot2)
p + scale_fill_manual(values = c("red","green"))

# change x-axis label
library(ggplot2)
p + xlab("Rate (events/Ma)")
}
}
