% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSSE.R
\name{plotHiSSE}
\alias{plotHiSSE}
\title{plotHiSSE}
\usage{
plotHiSSE(x)
}
\arguments{
\item{x}{a data frame containing columns "value", "rate", "hidden_state", "observed_state"}
}
\value{
a ggplot object
}
\description{
plotHiSSE
}
\examples{
hisse_file <- system.file("extdata", "sse/primates_HiSSE_2.log", package="RevGadgets")

tr <- readTrace(hisse_file)[[1]]

pdata <- processSSE(tr)
p <- plotHiSSE(pdata)
}
