% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMuSSE.R
\name{plotMuSSE}
\alias{plotMuSSE}
\title{plotMuSSE}
\usage{
plotMuSSE(rates)
}
\arguments{
\item{rates}{(data.frame; no default) a data frame containing columns "value", "rate",
"hidden_state", "observed_state" (such as the output of processSSE())}
}
\value{
a ggplot object
}
\description{
plotMuSSE
}
\examples{
\dontrun{
bisse_file <- system.file("extdata", "sse/primates_BiSSE_activity_period.log", package="RevGadgets")

pdata <- processSSE(bisse_file)
p <- plotMuSSE(pdata);p

# change colors:
p + scale_fill_manual(values = c("red","green"))

# change x-axis label
library(ggplot2)
p + xlab("Rate (events/Ma)")
}
}
