% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivRates.R
\name{plotPopulationSize}
\alias{plotPopulationSize}
\title{Title}
\usage{
plotPopulationSize(df, plot_var = "size")
}
\arguments{
\item{df}{}

\item{plot_var}{}
}
\value{
a ggplot object
}
\description{
Title
}
