% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStochMaps.R
\name{plotStochMaps}
\alias{plotStochMaps}
\title{plotStochMaps}
\usage{
plotStochMaps(
  tree,
  maps,
  colors = "default",
  color_by = "prob",
  tree_layout = "rectangular",
  line_width = 1,
  tip_labels = TRUE,
  tip_labels_italics = FALSE,
  tip_labels_formatted = FALSE,
  tip_labels_remove_underscore = TRUE,
  tip_labels_color = "black",
  tip_labels_size = 3,
  tip_labels_offset = 0,
  timeline = FALSE,
  geo_units = list("epochs", "periods"),
  geo = timeline,
  time_bars = timeline,
  label_sampled_ancs = FALSE,
  ...
)
}
\arguments{
\item{tree}{(treedata object; none) Output of readTrees() function
containing tree.}

\item{maps}{(dataframe; no default) Dataframe with processed maps,
as in the output of processStochMaps()}

\item{colors}{(named character vector; no default) Named character vector
where items are colors and names are the corresponding states.}

\item{color_by}{(character string; "prob") How to color the branches. 
Options are "MAP" for assigning color by the MAP state, or "prob" for 
assigning color as the weighted average of states based on the posterior
probabilities.}

\item{tree_layout}{(character; "rectangular") Tree shape layout, passed
to ggtree(). Options are 'rectangular', 'cladogram', 'slanted', 'ellipse',
'roundrect', 'fan', 'circular', 'inward_circular', 'radial', 'equal_angle',
'daylight', or 'ape'.}

\item{line_width}{(numeric; 1) Change line width for branches}

\item{tip_labels}{(logical; TRUE) Plot tip labels?}

\item{tip_labels_italics}{(logical; FALSE) Plot tip labels in italics?}

\item{tip_labels_formatted}{(logical; FALSE) Do the tip labels contain
manually added formatting information? Will set parse = TRUE in geom_text()
and associated functions to interpret formatting. See ?plotmath for more.
Cannot be TRUE if tip_labels_italics = TRUE.}

\item{tip_labels_remove_underscore}{(logical; TRUE) Remove underscores in tip
labels?}

\item{tip_labels_color}{(character; "black") Color to plot tip labels, either
as a valid R color name or a valid hex code.}

\item{tip_labels_size}{(numeric; 3) Size of tip labels}

\item{tip_labels_offset}{(numeric; 1) Horizontal offset of tip labels from
tree.}

\item{timeline}{(logical; FALSE) Plot time tree with labeled x-axis with
timescale in MYA.}

\item{geo_units}{(list; list("epochs", "periods")) Which geological units to
include in the geo timescale. May be "periods", "epochs", "stages", "eons",
"eras", or a list of two of those units.}

\item{geo}{(logical; timeline) Add a geological timeline? Defaults to the
same as timeline.}

\item{time_bars}{(logical; timeline) Add vertical gray bars to indicate
geological timeline units if geo == TRUE or regular time intervals (in MYA)
if geo == FALSE.}

\item{label_sampled_ancs}{(logical; FALSE) Label any sampled ancestors?
Will inherent tip labels aesthetics for size and color.}

\item{...}{(various) Additional arguments passed to ggtree::ggtree().}
}
\value{
returns a single plot object.
}
\description{
plotStochMaps
}
\examples{

\donttest{

# Standard stochastic mapping example

# read a tree (REPLACE WITH DOWNLOADING EXAMPLE BEFORE PUBLISHING)
treefile <- system.file("extdata",
                        "stoch_map_test_tmp/tree.nexus",
                        package="RevGadgets")
                        
tree <- readTrees(treefile)[[1]][[1]]

# process samples
mapsfile <- system.file("extdata",
                        "stoch_map_test_tmp/maps.log",
                        package="RevGadgets")
                        
stoch_map_df <- processStochMaps(tree,
                                 mapsfile, 
                                 states = as.character(0:4), 
                                 burnin = 0.1)

plotStochMaps(tree = tree,
              maps = stoch_map_df,
              color_by = "MAP",
              colors = "default",
              tree_layout = "rectangular",
              tip_labels = FALSE)

}

}
