% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrace.R
\name{plotTrace}
\alias{plotTrace}
\title{Plot trace}
\usage{
plotTrace(trace, vars, colors = "Set1", plot_together = TRUE)
}
\arguments{
\item{trace}{(list of data frames; no default) Name of a list of data frames,
such as produced by readTrace(). If the readTrace() output
contains multiple traces (such as from multiple runs), summarizeTrace() will provide
summaries for each trace individually, as well as the combined trace.}

\item{vars}{(character or character vector; no default) The name of the variable(s)
to be summarized.}

\item{colors}{(character vector; "Set1") Vector of colors for plotting, containing
the name of a Brewer palette, hex codes, or R-approved color names. Default behavior
is to plot using Color Brewer's Set1 palette.}

\item{plot_together}{(logical; TRUE) Plot all variables together?}
}
\value{
summarizeTrace() returns a list of the length of provided variables. For quantitative
variables, it returns the mean and 95% credible interval. For discrete variables, it returns
the 95% credible set of states and their associated probabilities.
}
\description{
Plots the posterior distributions of variables from trace file.
}
\details{
Plots the posterior distributions of continuous variables from one or
multiple traces (as in, from multiple runs). If multiple traces are provided,
plotTrace() will plot each run independently as well as plot the combined output.
When multiple variables are listed, the default behavior is to overlay
the variables, unless the user specifies combine = FALSE. Note that for variables
with very different distributions, overlaying the plots may result in illegible
figures. In these cases, we recommend combine = FALSE.
}
\examples{

\dontrun{
# continuous character only example, one run
file <- system.file("extdata",
    "sub_models/primates_cytb_covariotide.p", package="RevGadgets")
one_trace <- readTrace(path = file)
trace_sum <- summarizeTrace(trace = one_trace,
                            vars = c("Likelihood",
                                     "lambda",
                                   "TL"))

# discrete character example, multiple runs
file1 <- system.file("extdata",
    "comp_method_disc/mkstates_run_1.txt", package="RevGadgets")
file2 <- system.file("extdata",
    "comp_method_disc/mkstates_run_2.txt", package="RevGadgets")

multi_trace <- readTrace(path = c(file1, file2))
trace_sum_multi <- summarizeTrace(trace = multi_trace,
                                  vars = c("end_680",
                                           "end_681",
                                           "end_682",
                                           "end_683"))

}

}
