% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBranchData.R
\name{processBranchData}
\alias{processBranchData}
\title{processBranchData}
\usage{
processBranchData(
  tree,
  df,
  burnin = 0.25,
  parnames = c("avg_lambda", "avg_mu", "num_shifts"),
  summary = "median"
)
}
\arguments{
\item{tree}{a phylogenetic tree in the treedata format}

\item{df}{data frame with branch specific data}

\item{burnin}{fraction of the markov-chain to discard}

\item{parnames}{defaults to c("avg_lambda", "avg_mu", "num_shifts")}

\item{summary}{function to summarize the continuous parameter. Typically mean or median}
}
\value{
a treedata file with attached branch-specific data
}
\description{
processBranchData
}
\examples{
\dontrun{
library(RevGadgets)
library(tidytree)
library(ggplot2)
library(ape)

treefile <- system.file("extdata", "bds/primates.tre", package="RevGadgets")
logfile <- system.file("extdata", "bds/primates_BDS_rates_truncated.p", package="RevGadgets")

branch_data <- read.table(logfile, header = TRUE, sep = "\t")
tree <- tidytree::as.treedata(read.tree(treefile))

annotated_tree <- processBranchData(tree, branch_data, summary = "median")

p <- ggtree(annotated_tree) +
  aes(colour = avg_lambda) +
  theme(legend.position=c(0.2,0.80), legend.background=element_blank()) +
  scale_color_continuous("Posterior median \nspeciation rate",
                         low="blue", high="green")
}
}
