% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStochMaps.R
\name{processStochMaps}
\alias{processStochMaps}
\title{processStochMaps}
\usage{
processStochMaps(
  tree,
  paths = NULL,
  simmap = NULL,
  state_labels,
  num_intervals = 1000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{tree}{(treedata object; none) Output of readTrees() function
containing tree.}

\item{paths}{(vector of character strings; no default) File path(s) to
stochastic map trace(s).}

\item{simmap}{(multiphylo; none) A multiphylo object with simmaps in 
phytools format.}

\item{state_labels}{(named character vector; NULL) Vector of new labels for 
states named with the current state labels in annotated tree file
(as characters). If unnamed, state labels will be kept the same as in annotated
tree file. If any names are duplicated, then those columns with duplicated
names will be combined (summed).}

\item{num_intervals}{(numeric; default 1001) The number of intervals
to divide the tree into.}

\item{verbose}{(logical; default TRUE) Print status of processing on screen.}

\item{...}{(various) Additional arguments passed to readTrace()}
}
\description{
processStochMaps
}
\examples{

\donttest{

# download the example dataset to working directory

tree_url <- 
"https://revbayes.github.io/tutorials/morph_ase/data/solitariness_ase_hrm.tree"
tree_dest_path <- "solitariness_ase_hrm.tree"
download.file(tree_url, tree_dest_path)

maps_url <- 
"https://revbayes.github.io/tutorials/morph_ase/data/solitariness_hrm_stoch_char_map.log"
maps_dest_path <- "solitariness_hrm_stoch_char_map.log"
download.file(maps_url, maps_dest_path)

# to run on your own data, change these to the paths to your data files
tree_file <- tree_dest_path
maps_file <- maps_dest_path

# read in tree
tree <- readTrees(tree_file)

# default, don't rename states
stoch_map_df <- processStochMaps(tree,
                                 maps_file, 
                                 state_labels = as.character(c(0:3)), 
                                 burnin = 0.1)
# rename states 
stoch_map_df_named <- processStochMaps(tree,
                                       maps_file, 
                                       state_labels = c("no - slow" = "0", "yes - slow" = "1",
                                                        "no - fast" = "2", "yes - fast" = "3"), 
                                       burnin = 0.1)

# rename and combine states
stoch_map_df_combined <- processStochMaps(tree, 
                                          maps_file, 
                                          state_labels = c("no" = "0", "yes" = "1",
                                                           "no" = "2", "yes" = "3"), 
                                          burnin = 0.1)

# remove files
# WARNING: only run for example dataset!
# otherwise you might delete your data!
file.remove(tree_dest_path, maps_dest_path)

}

}
