% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStochMaps.R
\name{processStochMaps}
\alias{processStochMaps}
\title{processStochMaps}
\usage{
processStochMaps(
  tree,
  paths = NULL,
  simmap = NULL,
  states,
  num_intervals = 1000,
  ...
)
}
\arguments{
\item{tree}{(treedata object; none) Output of readTrees() function
containing tree.}

\item{paths}{(vector of character strings; no default) File path(s) to
stochastic map trace(s).}

\item{simmap}{(multiphylo; none) A multiphylo object with simmaps in 
phytools format.}

\item{states}{(vector of character strings; no default) The character
states.}

\item{num_intervals}{(numeric; default 1001) The number of intervals
to divide the tree in to.}

\item{...}{(various) Additional arguments passed to readTrace()}
}
\description{
processStochMaps
}
