% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiversityOBDP.R
\name{readOBDP}
\alias{readOBDP}
\title{Read OBDP outputs}
\usage{
readOBDP(
  start_time_trace_file,
  popSize_distribution_matrices_file,
  trees_trace_file
)
}
\arguments{
\item{start_time_trace_file}{(character; no default)  Trace of the starting times along the MCMC chain.}

\item{popSize_distribution_matrices_file}{(character; no default)  Kt matrices computed with `fnInferAncestralPopSize` in RevBayes.}

\item{trees_trace_file}{(character; no default)  Trace of the trees.}
}
\value{
A data.frame
}
\description{
Reads and formats the outputs of an analysis with the Occurrence Birth Death Process (MCMC parameter 
inference + diversity estimation)
}
