% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrace.R
\name{readTrace}
\alias{readTrace}
\title{Read trace}
\usage{
readTrace(
  paths,
  format = "simple",
  delim = "\\t",
  burnin = 0.1,
  check.names = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{paths}{(character vector) File path(s) to trace file(s).}

\item{format}{(character) Indicates type of MCMC trace. Options: "simple", "complex", "json".}

\item{delim}{(character) Delimiter of file (default: "\t").}

\item{burnin}{(numeric) Fraction of generations to discard (if < 1) or number of generations (if >= 1) (default: 0.1).}

\item{check.names}{(logical) Indicates if column names should be checked and replaced (default: FALSE).}

<<<<<<< HEAD
\item{verbose}{(logical; default = TRUE) Print status of reading traces 
to screen}

\item{...}{(various) Additional arguments passed to utils::read.table().}
=======
\item{verbose}{(logical) Whether to display progress messages (default: TRUE).}

\item{...}{(additional arguments) Passed to read.table() for file reading.}
>>>>>>> mahesh
}
\value{
List of data frames (of length 1 if only 1 log file provided).
}
\description{
Reads in MCMC log files and processes them according to specified parameters.
}
\examples{
\dontrun{
# Example usage:
file <- "../simple/part_run_1.log"
parsed_df <- readTrace(file, format = "simple")
View(parsed_df)  # View the parsed data frame
}

}
