% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrees.R
\name{readTrees}
\alias{readTrees}
\title{Read trees}
\usage{
readTrees(path, format)
}
\arguments{
\item{path}{(single character string; no default) File path to tree.}

\item{format}{(single character string; nexus) Format of tree file: nexus or newick}
}
\value{
Object of type multiPhylo, with length one if only one tree provided
}
\description{
Reads in a tree file containing one or multiple trees
}
\details{
Reads in a tree file in either nexus or newick format, and containing a single tree
or multiple trees (as in the results of a Bayesian analysis).
}
\examples{

\dontrun{
file <- system.file("extdata",
    "sub_models/primates_cytb_covariotide_MAP.tre", package="RevGadgets")
single_tree <- readTrees(path = file, format = "nexus")

file <- system.file("extdata",
    "sub_models/primates_cytb_covariotide.trees", package="RevGadgets")
multi_trees <- readTrees(path = file, format = "newick")
}
}
