% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrees.R
\name{readTrees}
\alias{readTrees}
\title{Read trees}
\usage{
readTrees(
  paths,
  tree_name = "psi",
  burnin = 0,
  n_cores = 1L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{paths}{(vector of character strings; no default) File path(s) to tree(s).}

\item{tree_name}{(character string; default psi) Name of the tree variable.}

\item{burnin}{(single numeric value; default = 0.1) Fraction of generations to
discard (if value provided is between 0 and 1) or number of generations (if
value provided is greater than 1).}

\item{n_cores}{(integer; default 1) Number of cores for parallelizing.}

\item{verbose}{(logical; default true) Display a status bar?}
}
\value{
A list (across runs) of lists (across samples) of treedata objects.
}
\description{
Reads in a tree file containing one or multiple trees
}
\details{
Reads in a tree file in either nexus or newick format, and containing a single tree
or multiple trees (as in the results of a Bayesian analysis).
}
\examples{


file <- system.file("extdata",
    "sub_models/primates_cytb_covariotide_MAP.tre", package="RevGadgets")
tree_single <- readTrees(paths = file)


}
