% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTrace.R
\name{summarizeTrace}
\alias{summarizeTrace}
\title{Summarize trace}
\usage{
summarizeTrace(trace, vars)
}
\arguments{
\item{trace}{(list of data frames; no default) Name of a list of data frames,
such as produced by readTrace(). If the readTrace() output
contains multiple traces (such as from multiple runs), summarizeTrace() will provide
summaries for each trace individually, as well as the combined trace.}

\item{vars}{(character or character vector; no default) The name of the variable(s)
to be summarized.}
}
\value{
summarizeTrace() returns a list of the length of provided variables. For quantitative
variables, it returns the mean and 95% credible interval. For discrete variables, it returns
the 95% credible set of states and their associated probabilities.
}
\description{
Summarizes trace file(s) that have been read into memory
}
\details{
Summarizes a trace file for continuous or discrete characters by
computing the mean and 95\% credible interval for quantitative
character and the 95\% credible set for discrete characters.
}
\examples{

\dontrun{
# continuous character only example, one run
file <- system.file("extdata",
    "sub_models/primates_cytb_covariotide.p", package="RevGadgets")
one_trace <- readTrace(path = file)
trace_sum <- summarizeTrace(trace = one_trace,
                            vars = c("pi[1]","pi[2]","pi[3]","pi[4]"))

# discrete character example, multiple runs
file1 <- system.file("extdata",
    "comp_method_disc/mkstates_run_1.txt", package="RevGadgets")
file2 <- system.file("extdata",
    "comp_method_disc/mkstates_run_2.txt", package="RevGadgets")

multi_trace <- readTrace(path = c(file1, file2))
trace_sum_multi <- summarizeTrace(trace = multi_trace,
                                  vars = c("end_680",
                                           "end_681",
                                           "end_682",
                                           "end_683"))

}

}
